diploma_spark/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
openai==1.13.3
anthropic==0.23.1
google-generativeai==0.5.2
mistralai==0.4.2
httpx==0.27.0
python-dotenv==1.0.1
reportlab==4.1.0
itsdangerous==2.1.2
aiofiles==23.2.1
